Add Multitask tutorial

+++ metal/multitask/mt_classifier.py
@@ -11,7 +11,7 @@ class MTClassifier(Classifier):
     The main contribution of children classes will be an implementation of the
     predict_proba() method. The relationships between the six predict/score
     functions are as follows:
-    score 		    	score_task
+        score 		    	score_task
             |			         |
         predict 	     	predict_task
             |	    (default)    |
@@ -109,7 +109,7 @@ class MTClassifier(Classifier):
         Y_ph = []
         for Y_tp in Y_p:
             Y_tph = self._break_ties(Y_tp, break_ties)
-            Y_ph.append(Y_tph)
+            Y_ph.append(Y_tph.astype(np.int))
 
         return Y_ph
